ref(core): Expose `prepareEvent` util function
