+++ packages/core/src/index.ts
@@ -27,6 +27,7 @@ export { createTransport } from './transports/base';
 export { SDK_VERSION } from './version';
 export { getIntegrationsToSetup } from './integration';
 export { FunctionToString, InboundFilters } from './integrations';
+export { prepareEvent } from './utils/prepareEvent';
 
 import * as Integrations from './integrations';
